debug en cours

+++ vm/src/instruction/xor.c
@@ -6,7 +6,7 @@
 /*   By: rostroh <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 13:25:27 by rostroh           #+#    #+#             */
-/*   Updated: 2017/12/13 23:01:55 by rostroh          ###   ########.fr       */
+/*   Updated: 2017/12/18 11:07:52 by rostroh          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void		ft_xor(t_inf inf, t_process *pros)
 {
-	unsigned int	res;
+	int	res;
 
 //	ft_printf(C_BLU"xor je t'envois au combat\t"FC_ALL);
 	if (inf.typ[0] == 1)
@@ -24,6 +24,6 @@ void		ft_xor(t_inf inf, t_process *pros)
 	res = inf.val[0] ^ inf.val[1];
 	pros->carry = (res == 0) ? 1 : 0;
 	reg_write(pros, (unsigned int)res, inf.val[2], REG_SIZE);
-	pros->val[inf.val[2]] = (unsigned int)res;
+	pros->val[inf.val[2]] = res;
 	//inf.val[2] = (int)res;
 }
